Set lib to "es5".
This is an optimization to speed up compilation because the default
for es5 as of today include es5, dom and scripthost. See --lib in
[typescript compiler-options](https://www.typescriptlang.org/docs/handbook/compiler-options.html).

For this tiny project the compile went from 1.6secs to 1.3secs, so
not really necessary but it seems that specifying target and lib is
a good idea.